add functional test with new map saved object
