+++ x-pack/test/functional/apps/maps/reporting/baseline/example_map_report.png
